Fix observable issues
