+++ src/Sidekick.Domain/Settings/CustomChatSetting.cs
@@ -6,36 +6,10 @@ using System.Text;
 
 namespace Sidekick.Domain.Settings
 {
-    public class CustomChatSetting : INotifyPropertyChanged
+    public class CustomChatSetting
     {
-        public event PropertyChangedEventHandler PropertyChanged;
+        public string Key { get; set; }
 
-        private string key;
-        private string chatCommand;
-
-        public string Key
-        {
-            get { return key; }
-            set
-            {
-                key = value;
-                OnPropertyChanged();
-            }
-        }
-
-        public string ChatCommand
-        {
-            get { return chatCommand; }
-            set
-            {
-                chatCommand = value;
-                OnPropertyChanged();
-            }
-        }
-
-        protected void OnPropertyChanged([CallerMemberName] string name = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-        }
+        public string ChatCommand { get; set; }
     }
 }
