+++ src/Sidekick.Presentation.Wpf/Settings/CustomChatModel.cs
@@ -1,0 +1,44 @@
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+using Sidekick.Domain.Settings;
+
+namespace Sidekick.Presentation.Wpf.Settings
+{
+    public class CustomChatModel : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        private string key;
+        private string chatCommand;
+
+        public string Key
+        {
+            get { return key; }
+            set
+            {
+                key = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public string ChatCommand
+        {
+            get { return chatCommand; }
+            set
+            {
+                chatCommand = value;
+                OnPropertyChanged();
+            }
+        }
+
+        protected void OnPropertyChanged([CallerMemberName] string name = null)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
+        }
+    }
+}
