+++ src/Sidekick.Presentation.Wpf/Settings/SettingsViewModel.cs
@@ -55,6 +55,9 @@ namespace Sidekick.Presentation.Wpf.Settings
             uiLanguageProvider.AvailableLanguages.ForEach(x => UILanguageOptions.Add(x.NativeName.First().ToString().ToUpper() + x.NativeName[1..], x.Name));
             gameLanguageProvider.AvailableLanguages.ForEach(x => ParserLanguageOptions.Add(x.Name, x.LanguageCode));
 
+            foreach (var setting in Custom_Chat_Settings)
+                CustomChatSettings.Add(new CustomChatModel { ChatCommand = setting.ChatCommand, Key = setting.Key });
+
             keybindsProvider.OnKeyDown += NativeKeyboard_OnKeyDown;
         }
 
@@ -74,9 +77,11 @@ namespace Sidekick.Presentation.Wpf.Settings
 
         public ObservableDictionary<string, string> ParserLanguageOptions { get; private set; } = new ObservableDictionary<string, string>();
 
+        public ObservableCollection<CustomChatModel> CustomChatSettings { get; private set; } = new ObservableCollection<CustomChatModel>();
+
         public string CurrentKey { get; set; }
 
-        public CustomChatSetting CurrentCustomChat { get; set; }
+        public CustomChatModel CurrentCustomChat { get; set; }
 
         public bool SettingCustom { get; set; }
 
@@ -144,7 +149,7 @@ namespace Sidekick.Presentation.Wpf.Settings
 
         public WikiSetting Wiki_Preferred { get; set; }
 
-        public ObservableCollection<CustomChatSetting> Custom_Chat_Settings { get; set; } = new ObservableCollection<CustomChatSetting>();
+        public List<CustomChatSetting> Custom_Chat_Settings { get; set; } = new List<CustomChatSetting>();
 
         #endregion
 
@@ -158,6 +163,10 @@ namespace Sidekick.Presentation.Wpf.Settings
         {
             var leagueHasChanged = LeagueId != sidekickSettings.LeagueId;
             var languageHasChanged = gameLanguageProvider.Current.LanguageCode != Language_Parser;
+
+            Custom_Chat_Settings.Clear();
+            foreach (var setting in CustomChatSettings)
+                Custom_Chat_Settings.Add(new CustomChatSetting { ChatCommand = setting.ChatCommand, Key = setting.Key });
 
             uiLanguageProvider.SetLanguage(Language_UI);
             await mediator.Send(new SetGameLanguageCommand(Language_Parser));
@@ -198,34 +207,32 @@ namespace Sidekick.Presentation.Wpf.Settings
                 SettingCustom = false;
                 return true;
             }
-            else
+
+            if (string.IsNullOrEmpty(CurrentKey))
             {
-                if (string.IsNullOrEmpty(CurrentKey))
-                {
-                    return false;
-                }
-
-                if (input == "Escape")
-                {
-                    CurrentKey = null;
-                    return true;
-                }
-
-                if (!IsKeybindUsed(input, CurrentKey))
-                {
-                    var property = GetType()
-                        .GetProperties()
-                        .FirstOrDefault(x => x.Name == CurrentKey);
-
-                    if (property != default)
-                    {
-                        property.SetValue(this, input);
-                    }
-                }
-
+                return false;
+            }
+            
+            if (input == "Escape")
+            {
                 CurrentKey = null;
                 return true;
             }
+            
+            if (!IsKeybindUsed(input, CurrentKey))
+            {
+                var property = GetType()
+                    .GetProperties()
+                    .FirstOrDefault(x => x.Name == CurrentKey);
+            
+                if (property != default)
+                {
+                    property.SetValue(this, input);
+                }
+            }
+            
+            CurrentKey = null;
+            return true;
         }
 
         public void Clear(string key)
@@ -270,9 +277,9 @@ namespace Sidekick.Presentation.Wpf.Settings
 
         public void NewCommand()
         {
-            if (!Custom_Chat_Settings.Any(x => x.ChatCommand == "New Command"))
+            if (!CustomChatSettings.Any(x => x.ChatCommand == "New Command"))
             {
-                Custom_Chat_Settings.Add(new CustomChatSetting { ChatCommand = "New Command", Key = "" });
+                CustomChatSettings.Add(new CustomChatModel { ChatCommand = "New Command", Key = "" });  
             }
         }
